+++ app/Http/Requests/DiaryRequest.php
@@ -28,7 +28,6 @@ class DiaryRequest extends FormRequest
         'tired'=>'required',
         'drink'=>'required',
         'hangover'=>'required',
-        
         ];
     }
